+++ Support/Extensions.cs
@@ -92,15 +92,14 @@ namespace Platform.Support
             return Helpers.IsNull<T>(value, replacement);
         }
 
-        public static T CType<T>(this object obj)
-        {
-            return (T)Helpers.CType<T>(obj);
-        }
+        //public static T CType<T>(this object obj)
+        //{
+        //    return (T)Helpers.CType<T>(obj);
+        //}
 
         #endregion
 
         #region Enums
-
 
 #if (!PORTABLE)
 
@@ -187,6 +186,41 @@ namespace Platform.Support
         {
             return Regex.Matches(text, ".{1," + lineLength + "}").Cast<Match>().Select(m => m.Value).ToArray();
         }
+
+        #endregion
+
+        #region DateTimes
+
+        public static string ISO8601(this System.DateTime date)
+        {
+            return Helpers.ISO8601(date);
+        }
+
+        public static bool IsMin(this System.DateTime date)
+        {
+            return date.Date == DateTime.MinValue.Date;
+        }
+
+        public static bool IsNow(this System.DateTime date)
+        {
+            return date.Date == DateTime.Now.Date;
+        }
+
+        //		public static System.DateTime IsNull (this System.DateTime? value, System.DateTime replacement)
+        //		{
+        //			if (value != null && value.GetType () == typeof(DateTime)) {
+        //				if (Convert.ToDateTime (value).ToString ("d") != Convert.ToDateTime (null).ToString ("d")) {
+        //					return value;
+        //				}
+        //			}
+        //			return replacement;
+        //		}
+        //
+        //		public static System.DateTime IsNull (this System.DateTime? value)
+        //		{
+        //			System.DateTime replacement = System.DateTime.MinValue;
+        //			return IsNull (value, replacement);
+        //		}
 
         #endregion