+++ Support/Reflection/Extensions.cs
@@ -68,5 +68,17 @@ namespace Platform.Support.Reflection
             return file.Directory.FullName;
         }
 
+
+
+        //public static AssemblyInfo GetAssemblyInfo(this object source)
+        //{
+        //    return new AssemblyInfo(source.GetType());
+        //}
+
+        //public static AssemblyInfo GetAssemblyInfo(this Type source)
+        //{
+        //    return new AssemblyInfo(source);
+        //}
+
     }
 }